requirements.txt
ccxt>=4.3.75
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
